AI-Model/Chat-models--main/requirements.txt
streamlit
scikit-learn
langchain
langchain-groq
langchain-google-genai
python-dotenv
